Fix Godot project & Framework+ui initialize and added GameManager.cs to godot Sample
