+++ Godot/GodotSample/project.godot
@@ -11,7 +11,7 @@ config_version=5
 [application]
 
 config/name="GodotSample"
-run/main_scene="res://ui.tscn"
+run/main_scene="res://main.tscn"
 config/features=PackedStringArray("4.1", "C#", "Forward Plus")
 config/icon="res://icon.svg"
